+++ example/lib/example.dart
@@ -1,6 +1,6 @@
 import 'dart:convert';
 
-import 'src/model/composition_test.dart';
+import 'src/model/data_composition_test.dart';
 import 'src/model/date_test.dart';
 import 'src/model/list_test.dart';
 import 'src/model/map_test.dart';