+++ lib/src/builder/entity_serializer_builder.dart
@@ -1,7 +1,9 @@
 import 'dart:async';
+import 'dart:io';
 
 import 'package:build/build.dart';
 import 'package:logging/logging.dart';
+import 'package:path/path.dart';
 import 'package:xml/xml.dart';
 
 import '../../entity_serializer.dart';
@@ -18,11 +20,15 @@ class EntitySerializerBuilder implements Builder {
   FutureOr<void> build(BuildStep buildStep) async {
     _dumpConfig();
     final outPath = buildStep.allowedOutputs.first.path;
+    final inPath = dirname(buildStep.inputId.path);
     final wrapper = GeneratorWrapper(
-        await buildStep.readAsString(buildStep.inputId), outPath);
+        await buildStep.readAsString(buildStep.inputId), inPath, outPath);
 
     final output = AssetId(buildStep.inputId.package, outPath);
-    return buildStep.writeAsString(output, wrapper.process());
+    final resultAsStr = wrapper.process();
+    if (resultAsStr.isNotEmpty) {
+      return buildStep.writeAsString(output, resultAsStr);
+    }
   }
 
   String get _getInputFolder => options.config["input_folder"] ?? "data_model";
@@ -48,21 +54,30 @@ class EntitySerializerBuilder implements Builder {
 
 class GeneratorWrapper {
   final String xmlString;
+  final String inPath;
   final String outPath;
   String? _resultString;
 
-  GeneratorWrapper(this.xmlString, this.outPath);
+  GeneratorWrapper(this.xmlString, this.inPath, this.outPath);
+
+  String _readFile(String filename) {
+    final filePath = join(inPath, filename);
+    return File(filePath).readAsStringSync();
+  }
 
   String process() {
     final document = XmlDocument.parse(xmlString);
-    final modelsGenerator = ModelGenerator(from: document);
-    final outputGenerator = OutputGenerator.forBuilder(
-      models: modelsGenerator,
-      outputConsumer: _consume,
-      outputPath: outPath,
-    );
-    outputGenerator.generateOutput();
-    return _resultString!;
+    Model model = Model();
+    ModelGenerator(from: document, model: model, readFileFunc: _readFile);
+    if (model.isNotEmpty) {
+      final outputGenerator = OutputGenerator.forBuilder(
+        models: model,
+        outputConsumer: _consume,
+        outputPath: outPath,
+      );
+      outputGenerator.generateOutput();
+    }
+    return _resultString ?? "";
   }
 
   void _consume(String data) {