+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.0.4
+- Feature: compose single model from multiple xml files
+
 ## 0.0.3
 - Bugfix: optional values generation fix
 - Feature: Serialization field alias per serializer specialization